Updated DOMStatus sector. -New I3DaqRawDOMStatus class to hold Real Daq per-run configurations. -Typo in I3Calibration serialization method -Smaller virtual interface for I3RawDOMStatus, removed MC specific stuff 	(N sample stuff) -Added static_cast for int to double conversion
git-svn-id: http://code.icecube.wisc.edu/svn/projects/dataclasses/trunk@12502 16731396-06f5-0310-8873-f7f720988828

+++ public/dataclasses/I3MCRawDOMStatus.h
@@ -128,9 +128,9 @@ public:
     virtual void SetDACTriggerBias(int bias, int chip)
 	{
 	    if ( chip == 0 )
-		dacTriggerBias0_ = bias;
+		dacTriggerBias0_ = static_cast<double>(bias);
 	    else if ( chip == 1 )
-		dacTriggerBias1_ = bias;
+		dacTriggerBias1_ = static_cast<double>(bias);
 	    else 
 		log_fatal("Bad chip ID in I3MCRawDOMStatus::SetDACTriggerBias");
 	};
